+++ underscoreLoader.js
@@ -11,7 +11,7 @@ const defaults = {
   minifierOptions: {
     removeComments: true,
     collapseWhitespace: true,
-    conservativeCollapse: true
+    collapseInlineTagWhitespace: true
   },
   originalSource: true,
   templateOptions: {}